After discussion with Philippe, remove ReportedSiteName from ProbeSummary entirely.
This involves re-writing the "odd site name" query report.

Also, completely overhaul the method for upgrading the auxiliary DB
items like stored procedures and summary tables since when upgrading
from an old version of the DB a too-early application of (eg) a modern
trigger could fail. So, keep track of the last DB version to require an
update of a particular item and only apply it after a successful upgrade
of the main DB schema.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@866 c85a5f45-a5f3-44fc-b02c-7891c381c519