More statistics when operating in verbose mode.
Only set Grid=Local if we're sure that the ClassAd attribute should have
been added by the JobManager (assuming it passed through same) but was
not (indicating a local job).

Improbe logic testing whether we can rely on the check for the
JobManager in the case that the site is running Condor for managed fork
and PBS for standard batch jobs.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@2120 c85a5f45-a5f3-44fc-b02c-7891c381c519