Fix problem starting checksum upgrade thread in CollectorService.
Housekeeping can be disabled (as opposed to stopped). This is detected
by the listener threads, which then proceed to accept old records (if
the housekeeping were merely stopped, old records will still be rejected).

Administration page reflects more detailed checksum upgrade and
housekeeping status, and the ability to disable the housekeeping and
accept old records.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@2258 c85a5f45-a5f3-44fc-b02c-7891c381c519