Overhaul NodeSummary update to improve efficiency.
NodeSummary is now updated when del_JUR_from_summary.

Unrelated to above: create NodeSummary table in *all* instances, not
just when wantNodeSummary is set. However, only update NodeSummary when
wantNodeSummary is set. This is to work around an occasional behavior
whereby MySQL will complain that NodeSummary does not exist *even
though* the conditional is verified to be correct and operating. In
other words, during testing we discovered that MySQL would complain
about code *that is NOT executed*. Tests have been exhausted and we have
been unable to ascertain exactly what conditions combine to make MySQL
do this (not able to reproduce on gratia-vm02, for example), although it
appears that the primary trigger is having a ps-accounting collector in
the same DB create a NodeSummary table via hibernate. Everything else is
speculation. By creating the NodeSummary table in all circumstances
however, we have bypassed the problem at the expense of having a
(currently) vestigial table present on non-ps-accounting collectors.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@2300 c85a5f45-a5f3-44fc-b02c-7891c381c519