Today's commit-and-run:
Satisfied requirements on transfer job-level and summary-level data.

New tables:

* MasterTransferSummary (standalone)
* TransferDetails (linked to TDCorr via TransferDetailsId)
* TDCorr (linked to TransferDetails via TransferDetailsId and to JobUsageRecord and friends via dbid).

Several interesting workarounds required:

1. ListenerThread is more careful about checking current.getDate() since
it may not always get a good one from the record.

2. add_JUR_to_summary now has an extra argument and is now called by
*two* triggers. According to the rules of hibernate, collections are
committed *last*, meaning that a trigger on JobUsageRecord_Meta can't
update the TransferSummary because TDCorr and TransferDetails haven't
been written yet. So, we have to refrain from writing the summary entry
on JobUsageRecord_Meta insert if it's a transfer record and instead wait
until the routine is called from a trigger on TDCorr insert.

Other notes:

1. Housekeeping will need to be updated with instructions on cleaning up
the two new tables.

2. JobUsageRecord has a more intelligent getDate() method.

3. I hate hibernate.

4. I hate kludgy hibernate / JDBC / DB trigger hybrids even more.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@2434 c85a5f45-a5f3-44fc-b02c-7891c381c519