Add ability to call in "human-readable mode"
Fix a broken regex.

Improve circumstances when manual-root message is printed.

Better message when PBS / LSF tests are not run.

Add "success" messages to detailsData.

Add help option and usage message.

Better error messages if probe is not specified.

Only check crontab if we know it's a probe that needs one.

git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@3226 c85a5f45-a5f3-44fc-b02c-7891c381c519