Increase the default batch size for row deletion from 200 to 1000. Change the default fault batch size for _Xml row deletion to 100000.
Introduce the 2 customizable parameters:

service.lifetimeManagement.BatchSize and service.lifetimeManagement.XmlBatchSize

Rework the row deletion from RecordType_Xml.

a. We no longer keep the row when the EndTime is greater than the ServerDate (and EndTime was greater than the cutoff date).
b. We no longer use a temporary table.
c. Improve the 'limited' delete query by using a dbid range to restrict the searches.
c.1. We find the maximum plausible dbid by searching in RecordType_Meta for the dbid clause to or equal to the limit date
c.2. We find the minimum plausible dbid by searching in RecordType_Xml for the first dbid with ExtraXml==""
c.3. the delete query where clause is:
   ExtraXml = "" and ServerDate < :dateLimit and :mindbid <= X.dbid and X.dbid < :maxdbid

On gr8x0 the current (typical rate) are (for a Transfer record database):
   JobUsageRecord_Xml 10000 rows per second (so for a typical transaction time of 10s).
   JobUsageRecord_Meta  120 rows per second (so for a typical transaction time of  9s).
   The other deletion are more than 1000 rows per seconds such that the total time is completely dominated by _Meta.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@4065 c85a5f45-a5f3-44fc-b02c-7891c381c519