This was created due to a problem with PBS probes which were creating probe records with ridiculously high CPU times under some circumstances. No one could find just what was causing the bug, so a patch was made to the PBS probe to quarantine anything higher that 1000 percent efficiency.   https://jira.opensciencegrid.org/browse/SOFTWARE-1032
This is set to go into OSG 3.1.19 as of 6/21/13 anyway.

When sites actually upgrade to this is an unknown.

The jur-efficiency-cleanup.sh was created to identify sites/probes causing
these distorted CPUs.  The usage function in the script describes the options
required.... run with a -h option.

When run, the script will create a directory for the database being accessed,
The gratia database is currently the only one of importance.  It will contain
the following files:

./gratia/analysis.log
   This shows all the queries and results performed.

./grata/del_JUR_from_summary.PROBE_NAME
   This contains the sql commands to all the del_JUR_from_summary for each dbid
    with a high CPU.

./grata/update_jur_records.PROBE_NAME
   This contains the sql commands to zero out the CPU and Wall times on the
   JobUsageRecords for each dbid with high CPU.




git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@5054 c85a5f45-a5f3-44fc-b02c-7891c381c519