JIRA ticket Gratia-124
This change is only to the del_JUR_from_summary procedure.
Added an update of the JobUsageRecord setting the respective values 
to zero based on the summary table updated.  The rationale for this is
to eliminate the chance of accidentally running the procedure twice and
inadvertantly doubling the up the decrement.
.
For MasterSummaryData table
----------------------------
Njobs
WallDuration
CpuSystemDuration
CpuUserDuration
.
For MasterTransferSummary table
--------------------------------
Njobs
TransferSize
TransferDuration
.
This makes the correction process for removing bad JobUsageRecord
data from the summary tables a simple one-step process.  If one 
inadvertantly runs this against the wrong dbid, the trace table has the
original values in the message it outputs in the sqlQuery column thereby
allowing for the JobUsageRecord to be repopulate and the add_JUR procedure
run to restore it.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@5132 c85a5f45-a5f3-44fc-b02c-7891c381c519